Restructure for Poetry package
