+++ pyproject.toml
@@ -1,0 +1,16 @@
+[tool.poetry]
+name = "database-stream"
+version = "0.1.0"
+description = "Streaming text through database tables"
+authors = ["Your Name <you@example.com>"]
+packages = [{include = "database_stream", from = "src"}]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+
+[tool.poetry.dev-dependencies]
+pytest = "*"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
