+++ src/database_stream/__init__.py
@@ -1,0 +1,6 @@
+"""Core classes for streaming text to and from database tables."""
+
+from .db_stream import DatabaseStream
+from .sqlite_db_stream import SqliteDatabaseStream
+
+__all__ = ["DatabaseStream", "SqliteDatabaseStream"]
